feat: Refactor lib, add support for clone of GossipDiscoveryReceiver
